Move Opinfo testing to a separate folder. (#4731)
This PR moves the opinfo tests into a separate directory and creates
`opinfo_utils.py`.

Details:
* Moves `test_cat_qwen2_v2` to `tests/python/test_python_frontend.py`.
It isn't supposed to be in `tests/test_ops.py`.
