+++ tests/opinfo/README.md
@@ -8,10 +8,10 @@
 
 ## Usage
 
-* Run tests: `pytest python_tests/test_ops.py`
-* Filter tests with `-k` option: `pytest python_tests/test_ops.py -k var_mean`
-* Show all possible tests: `pytest python_tests/test_ops.py --collect-only`
-* Filter all possible tests with `-k` option: `pytest python_tests/test_ops.py --collect-only -k var_mean`
+* Run tests: `pytest tests/opinfo/test_ops.py`
+* Filter tests with `-k` option: `pytest tests/opinfo/test_ops.py -k var_mean`
+* Show all possible tests: `pytest tests/opinfo/test_ops.py --collect-only`
+* Filter all possible tests with `-k` option: `pytest tests/opinfo/test_ops.py --collect-only -k var_mean`
 
 ## Dependencies
 * `pytest`
