+++ tests/opinfo/test_ops.py
@@ -12,10 +12,12 @@ from opinfo_fusion_definitions import default_fd_fn, parse_inputs_fusion_definit
 from opinfo_framework import create_op_test, atexit_serde_create_op_test
 from opinfo_core import ReferenceType, OpInfo, SampleInput
 from opinfos import opinfos
-from nvfuser.testing.utils import (
+from opinfo_utils import (
     ArgumentType,
     is_tensor,
     requiresJAX,
+)
+from nvfuser.testing.utils import (
     check_captured_python_definition,
     check_cpp_translation,
     debug_serde,
@@ -23,7 +25,7 @@ from nvfuser.testing.utils import (
 )
 from typing import Callable
 
-from nvfuser import DataType, FusionCache, FusionDefinition
+from nvfuser import FusionCache, FusionDefinition
 from nvfuser.pytorch_utils import retry_on_oom_or_skip_test
 
 
@@ -288,198 +290,3 @@ def test_errors(op: OpInfo, dtype: torch.dtype):
             exception_type, match=_regex_escape_parenthesis(exception_regex)
         ):
             errors_test_fn(op, sample)
-
-
-@pytest.mark.skip("https://github.com/NVIDIA/Fuser/issues/3740")
-def test_cat_qwen2_v2():
-    def qwen2_cat_fusion_2(fd: FusionDefinition) -> None:
-        T0 = fd.define_tensor(
-            shape=[2048, 512],
-            contiguity=[True, True],
-            dtype=DataType.BFloat16,
-            is_cpu=False,
-            stride_order=[1, 0],
-        )
-        S1 = fd.define_scalar(None, dtype=DataType.Int)
-        S2 = fd.define_scalar(None, dtype=DataType.Int)
-        T3 = fd.define_tensor(
-            shape=[1, 4, 2048, 128],
-            contiguity=[None, True, True, True],
-            dtype=DataType.BFloat16,
-            is_cpu=False,
-            stride_order=[3, 1, 2, 0],
-        )
-        T4 = fd.define_tensor(
-            shape=[1, 2048, 128],
-            contiguity=[None, True, True],
-            dtype=DataType.BFloat16,
-            is_cpu=False,
-            stride_order=[2, 0, 1],
-        )
-        T5 = fd.define_tensor(
-            shape=[1, 2048, 128],
-            contiguity=[None, True, True],
-            dtype=DataType.BFloat16,
-            is_cpu=False,
-            stride_order=[2, 0, 1],
-        )
-        T6 = fd.define_tensor(
-            shape=[1, 28, 2048, 128],
-            contiguity=[None, True, True, True],
-            dtype=DataType.BFloat16,
-            is_cpu=False,
-            stride_order=[3, 1, 2, 0],
-        )
-        T7 = fd.define_tensor(
-            shape=[1, 2048, 512],
-            contiguity=[None, True, True],
-            dtype=DataType.BFloat16,
-            is_cpu=False,
-            stride_order=[2, 1, 0],
-        )
-        T12 = fd.ops.reshape(T0, new_shape=[1, 2048, 512])
-        T13 = fd.ops.cast(T12, dtype=DataType.Float)
-        S14 = fd.define_scalar(0.00000, dtype=DataType.Double)
-        S15 = fd.define_scalar(1.00000, dtype=DataType.Double)
-        S16 = fd.define_scalar(1, dtype=DataType.Int)
-        S17 = fd.define_scalar(2048, dtype=DataType.Int)
-        S18 = fd.define_scalar(512, dtype=DataType.Int)
-        T20 = fd.ops.uniform(
-            S14,
-            S15,
-            shape=[S16, S17, S18],
-            rng_seed=S2,
-            rng_offset=S1,
-            dtype=DataType.BFloat16,
-        )
-        S21 = fd.define_scalar(4.00000, dtype=DataType.Double)
-        T22 = fd.ops.mul(T13, S21)
-        S23 = fd.define_scalar(0.900000, dtype=DataType.Double)
-        T24 = fd.ops.lt(T20, S23)
-        T25 = fd.ops.cast(T24, dtype=DataType.Float)
-        T41 = fd.ops.slice(
-            T3,
-            start_indices=[0, 0, 0, 64],
-            end_indices=[1, 4, 2048, 128],
-            strides=[1, 1, 1, 1],
-            manual_normalization=0,
-        )
-        T42 = fd.ops.mul(T22, T25)
-        T43 = fd.ops.cast(T41, dtype=DataType.Float)
-        T44 = fd.ops.neg(T43)
-        T50 = fd.ops.broadcast_in_dim(
-            T4, shape=[1, 1, 2048, 128], broadcast_dims=[0, 2, 3]
-        )
-        T66 = fd.ops.slice(
-            T3,
-            start_indices=[0, 0, 0, 0],
-            end_indices=[1, 4, 2048, 64],
-            strides=[1, 1, 1, 1],
-            manual_normalization=0,
-        )
-        T67 = fd.ops.cast(T44, dtype=DataType.BFloat16)
-        T73 = fd.ops.broadcast_in_dim(
-            T5, shape=[1, 1, 2048, 128], broadcast_dims=[0, 2, 3]
-        )
-        T89 = fd.ops.slice(
-            T6,
-            start_indices=[0, 0, 0, 64],
-            end_indices=[1, 28, 2048, 128],
-            strides=[1, 1, 1, 1],
-            manual_normalization=0,
-        )
-        S90 = fd.define_scalar(1.11111, dtype=DataType.Double)
-        T91 = fd.ops.mul(T42, S90)
-        T97 = fd.ops.broadcast_in_dim(
-            T50, shape=[1, 4, 2048, 128], broadcast_dims=[0, 1, 2, 3]
-        )
-        T98 = fd.ops.cat([T67, T66], dim=-1, manual_padding=0)
-        T104 = fd.ops.broadcast_in_dim(
-            T73, shape=[1, 4, 2048, 128], broadcast_dims=[0, 1, 2, 3]
-        )
-        T105 = fd.ops.cast(T89, dtype=DataType.Float)
-        T106 = fd.ops.cast(T97, dtype=DataType.Float)
-        T107 = fd.ops.cast(T98, dtype=DataType.Float)
-        T108 = fd.ops.cast(T104, dtype=DataType.Float)
-        T109 = fd.ops.cast(T3, dtype=DataType.Float)
-        T110 = fd.ops.neg(T105)
-        T111 = fd.ops.cast(T7, dtype=DataType.Float)
-        T112 = fd.ops.mul(T107, T106)
-        T113 = fd.ops.mul(T109, T108)
-        T129 = fd.ops.slice(
-            T6,
-            start_indices=[0, 0, 0, 0],
-            end_indices=[1, 28, 2048, 64],
-            strides=[1, 1, 1, 1],
-            manual_normalization=0,
-        )
-        T130 = fd.ops.cast(T110, dtype=DataType.BFloat16)
-        T131 = fd.ops.add(T111, T91)
-        T137 = fd.ops.broadcast_in_dim(
-            T50, shape=[1, 28, 2048, 128], broadcast_dims=[0, 1, 2, 3]
-        )
-        T138 = fd.ops.cat([T130, T129], dim=-1, manual_padding=0)
-        T144 = fd.ops.broadcast_in_dim(
-            T73, shape=[1, 28, 2048, 128], broadcast_dims=[0, 1, 2, 3]
-        )
-        T145 = fd.ops.cast(T131, dtype=DataType.BFloat16)
-        T146 = fd.ops.cast(T137, dtype=DataType.Float)
-        T147 = fd.ops.cast(T138, dtype=DataType.Float)
-        T148 = fd.ops.cast(T144, dtype=DataType.Float)
-        T149 = fd.ops.cast(T6, dtype=DataType.Float)
-        T155 = fd.ops.reshape(T145, new_shape=[1, 2048, 4, 128])
-        T156 = fd.ops.add(T113, T112)
-        T157 = fd.ops.mul(T147, T146)
-        T158 = fd.ops.mul(T149, T148)
-        T159 = fd.ops.permute(T155, dims=[0, 2, 1, 3])
-        T160 = fd.ops.cast(T156, dtype=DataType.BFloat16)
-        T167 = fd.ops.broadcast_in_dim(
-            T159, shape=[1, 4, 1, 2048, 128], broadcast_dims=[0, 1, 3, 4]
-        )
-        T174 = fd.ops.broadcast_in_dim(
-            T160, shape=[1, 4, 1, 2048, 128], broadcast_dims=[0, 1, 3, 4]
-        )
-        T181 = fd.ops.broadcast_in_dim(
-            T167, shape=[1, 4, 7, 2048, 128], broadcast_dims=[0, 1, 2, 3, 4]
-        )
-        T188 = fd.ops.broadcast_in_dim(
-            T174, shape=[1, 4, 7, 2048, 128], broadcast_dims=[0, 1, 2, 3, 4]
-        )
-        T189 = fd.ops.add(T158, T157)
-        T195 = fd.ops.reshape(T181, new_shape=[1, 28, 2048, 128])
-        T201 = fd.ops.reshape(T188, new_shape=[1, 28, 2048, 128])
-        T202 = fd.ops.cast(T189, dtype=DataType.BFloat16)
-        T203 = fd.ops.stride_order(T195, stride_order=[3, 2, 1, 0])
-        T204 = fd.ops.stride_order(T201, stride_order=[3, 2, 1, 0])
-        T205 = fd.ops.stride_order(T202, stride_order=[3, 2, 1, 0])
-        fd.add_output(T159)
-        fd.add_output(T160)
-        fd.add_output(T203)
-        fd.add_output(T204)
-        fd.add_output(T205)
-
-    inputs = [
-        torch.testing.make_tensor((2048, 512), dtype=torch.bfloat16, device="cuda:0"),
-        25546,
-        1400552702872758,
-        torch.randn(1048576, dtype=torch.bfloat16, device="cuda:0").as_strided(
-            (1, 4, 2048, 128), (1048576, 128, 512, 1)
-        ),
-        torch.randn(262144, dtype=torch.bfloat16, device="cuda:0").as_strided(
-            (1, 2048, 128), (262144, 1, 2048)
-        ),
-        torch.randn(262144, dtype=torch.bfloat16, device="cuda:0").as_strided(
-            (1, 2048, 128), (262144, 1, 2048)
-        ),
-        torch.randn(7340032, dtype=torch.bfloat16, device="cuda:0").as_strided(
-            (1, 28, 2048, 128), (7340032, 128, 3584, 1)
-        ),
-        torch.testing.make_tensor(
-            (1, 2048, 512), dtype=torch.bfloat16, device="cuda:0"
-        ),
-    ]
-
-    with FusionDefinition() as fd:
-        qwen2_cat_fusion_2(fd)
-
-    fd.execute(inputs)
